7. api call 기반으로 패키지 분리
7. api call 기반으로 패키지 분리
