+++ moim/src/main/java/com/youth/moim/domain/Gender.java
@@ -1,0 +1,5 @@
+package com.youth.moim.domain;
+
+public enum Gender {
+    FEMALE, MALE
+}
