+++ moim/src/main/java/com/youth/moim/domain/UserReader.java
@@ -1,0 +1,7 @@
+package com.youth.moim.domain;
+
+import java.util.Optional;
+
+public interface UserReader {
+    Optional<User> findByEmail(String email);
+}
