+++ moim/src/main/java/com/youth/moim/domain/UserStore.java
@@ -1,0 +1,5 @@
+package com.youth.moim.domain;
+
+public interface UserStore {
+    void registerUser(User entity);
+}
