+++ moim/src/main/java/com/youth/moim/infrastructure/UserReaderImpl.java
@@ -1,0 +1,17 @@
+package com.youth.moim.infrastructure;
+
+import com.youth.moim.domain.User;
+import com.youth.moim.domain.UserReader;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+import java.util.Optional;
+
+@Component
+@RequiredArgsConstructor
+public class UserReaderImpl implements UserReader {
+    private final UserRepository userRepository;
+    public Optional<User> findByEmail(String email) {
+        return userRepository.findByEmail(email);
+    }
+}
