+++ moim/src/main/java/com/youth/moim/infrastructure/UserRepository.java
@@ -1,0 +1,14 @@
+package com.youth.moim.infrastructure;
+
+import com.youth.moim.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
