--- moim/src/test/java/com/youth/moim/presentation/user/Gender.java
@@ -1,5 +1,0 @@
-package com.youth.moim.presentation.user;
-
-public enum Gender {
-    FEMALE, MALE
-}
