--- moim/src/test/java/com/youth/moim/presentation/user/UserReader.java
@@ -1,15 +1,0 @@
-package com.youth.moim.presentation.user;
-
-import java.util.Optional;
-
-public class UserReader {
-    private final UserRepository userRepository;
-
-    public UserReader(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    public Optional<User> findByEmail(String email) {
-        return userRepository.findByEmail(email);
-    }
-}
